+++ .ci/slurm.sh
@@ -1,6 +1,0 @@
-#!/usr/bin/env bash
-
-/etc/init.d/slurmd start
-/etc/init.d/munge start
-/etc/init.d/slurmctld start
-