+++ src/neptune/ExtractSignatures.py
@@ -3,7 +3,7 @@
 """
 # =============================================================================
 
-Copyright Government of Canada 2015-2017
+Copyright Government of Canada 2015-2024
 
 Written by: Eric Marinier, Public Health Agency of Canada,
     National Microbiology Laboratory
@@ -78,11 +78,11 @@ import math
 import argparse
 import os
 
-from Utility import reverseComplement
-from Utility import buildReferences
-from Utility import estimateReferenceParameters
+from neptune.Utility import reverseComplement
+from neptune.Utility import buildReferences
+from neptune.Utility import estimateReferenceParameters
 
-import Signature
+import neptune.Signature as Signature
 
 from scipy.stats import norm
 
@@ -413,10 +413,10 @@ def calculateProbHBMM(GC):
 
     # P(X_M = Y_M)_H
     probHBMM = (
-        2 * math.pow(GC / (GC + 1), 2) +
-        math.pow((1 - GC) / (GC + 1), 2)) * (1 - GC) +\
-        (2 * math.pow((1 - GC) / (2 - GC), 2) +
-            math.pow(GC / (2 - GC), 2)) * (GC)
+        2 * math.pow(GC / (GC + 1), 2)
+        + math.pow((1 - GC) / (GC + 1), 2)) * (1 - GC) \
+        + (2 * math.pow((1 - GC) / (2 - GC), 2)
+           + math.pow(GC / (2 - GC), 2)) * (GC)
 
     return probHBMM
 
@@ -1013,7 +1013,7 @@ def parse(parameters):
 
     # --- k-mer Size ---
     if not os.path.isfile(parameters[KMERS]):
-            raise RuntimeError("ERROR: Could not open k-mer file.\n")
+        raise RuntimeError("ERROR: Could not open k-mer file.\n")
 
     kmerLocation = parameters[KMERS]
     kmerFile = open(kmerLocation, 'r')