+++ docs/news.md
@@ -1,6 +1,12 @@
-# News #
+# News
 
-## Version 1.2.5 ##
+## Version 2.0.0
+
+2024-10-21
+
+This release updates Neptune to Python3, removes DRMAA support, fixes a crash when no signatures are produced, and updates the installation process.
+
+## Version 1.2.5
 
 2017-05-04
 
@@ -12,7 +18,7 @@ Neptune version 1.2.5 has been released to GitHub, Bioconda, and Galaxy:
 
 Please see the [Change Log](CHANGELOG) for additional information.
 
-## Bioconda Installation ##
+## Bioconda Installation
 
 2017-04-07